# docs/swagger.yaml
definitions:
  dto.AuditLogListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.AuditLogResponse'
        type: array
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  dto.AuditLogResponse:
    properties:
      action:
        type: string
      created_at:
        type: string
      id:
        type: integer
      ip:
        type: string
      resource_id:
        type: string
      resource_type:
        type: string
      tenant_id:
        type: string
      user_agent:
        type: string
      user_id:
        type: string
    type: object
  dto.AuthResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.CreateTaskRequest:
    properties:
      description:
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - title
    type: object
  dto.LoginRequest:
    properties:
      captcha_code:
        type: string
      captcha_id:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - captcha_code
    - captcha_id
    - password
    - username
    type: object
  dto.LogoutRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.PasswordResetRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  dto.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.RegisterRequest:
    properties:
      captcha_code:
        type: string
      captcha_id:
        type: string
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - captcha_code
    - captcha_id
    - password
    - username
    type: object
  dto.ResetPasswordRequest:
    properties:
      new_password:
        minLength: 6
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
  dto.TaskListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.TaskResponse'
        type: array
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  dto.TaskResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      status:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  dto.UpdateProfileRequest:
    properties:
      avatar_url:
        type: string
      full_name:
        type: string
      phone:
        type: string
    type: object
  dto.UpdateTaskRequest:
    properties:
      description:
        type: string
      status:
        enum:
        - 0
        - 1
        type: integer
      title:
        maxLength: 200
        minLength: 1
        type: string
    type: object
  dto.UserListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  dto.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      role:
        type: string
      status:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /admin/audit-logs:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuditLogListResponse'
      security:
      - BearerAuth: []
      summary: 管理员查询审计日志
      tags:
      - admin
  /admin/tasks:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskListResponse'
      security:
      - BearerAuth: []
      summary: 管理员获取所有任务
      tags:
      - admin
  /admin/tasks/{id}:
    delete:
      parameters:
      - description: 任务 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 管理员强制删除任意任务
      tags:
      - admin
  /admin/users:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserListResponse'
      security:
      - BearerAuth: []
      summary: 管理员获取所有用户
      tags:
      - admin
  /admin/users/{id}:
    get:
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
      security:
      - BearerAuth: []
      summary: 管理员获取指定用户
      tags:
      - admin
  /admin/users/{id}/ban:
    put:
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 管理员封禁用户
      tags:
      - admin
  /admin/users/{id}/demote:
    put:
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 管理员降级用户
      tags:
      - admin
  /admin/users/{id}/promote:
    put:
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 管理员提升用户为 admin
      tags:
      - admin
  /admin/users/{id}/unban:
    put:
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 管理员解封用户
      tags:
      - admin
  /auth/captcha:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取验证码
      tags:
      - auth
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: 邮箱信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.PasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 发起密码重置
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
      summary: 登录
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh Token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 登出
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh Token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
      summary: 刷新 Token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 注册信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
      summary: 注册
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: 重置密码请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 重置密码
      tags:
      - auth
  /healthz:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 存活检查
      tags:
      - health
  /readyz:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 就绪检查（含 DB ping）
      tags:
      - health
  /tasks:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskListResponse'
      security:
      - BearerAuth: []
      summary: 获取当前用户任务列表
      tags:
      - tasks
    post:
      consumes:
      - application/json
      parameters:
      - description: 任务信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskResponse'
      security:
      - BearerAuth: []
      summary: 创建任务
      tags:
      - tasks
  /tasks/{id}:
    delete:
      parameters:
      - description: 任务 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除任务
      tags:
      - tasks
    get:
      parameters:
      - description: 任务 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskResponse'
      security:
      - BearerAuth: []
      summary: 获取任务详情
      tags:
      - tasks
    put:
      consumes:
      - application/json
      parameters:
      - description: 任务 ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskResponse'
      security:
      - BearerAuth: []
      summary: 更新任务
      tags:
      - tasks
  /upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 上传文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 上传文件
      tags:
      - upload
  /user/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
      security:
      - BearerAuth: []
      summary: 获取当前用户资料
      tags:
      - user
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
      security:
      - BearerAuth: []
      summary: 更新当前用户资料
      tags:
      - user
swagger: "2.0"